Move add_repeated_pattern to parser
